Move BulkEditing to the left
